Move wallet code to the crate
- Delete unneeded modules
